// apps/frontend/src/app/api/models/api-error-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ValidationErrorDto } from '../models/validation-error-dto';
export interface ApiErrorResponse {
  error?: string;
  message?: string;
  path?: string;
  requestId?: string;
  status?: number;
  timestamp?: string;
  validationErrors?: Array<ValidationErrorDto>;
}